refactor project to MVC structure
