+++ app.py
@@ -1,8 +1,9 @@
 from flask import Flask
-from database.database import db
-from database.models import User
+from models.database import db
+from models.user import User
 from os import environ
 from flask_login import LoginManager
+
 
 def create_app():
     app = Flask(__name__)
@@ -14,15 +15,15 @@ def create_app():
     app.config["SECRET_KEY"] = "abc"
 
     db.init_app(app)
-	
+
     login_manager = LoginManager()
     login_manager.init_app(app)
-	
+
     @login_manager.user_loader
     def loader_user(user_id):
         return User.query.get(user_id)
 
-    from routes import routes
+    from routes.routes import routes
     app.register_blueprint(routes)
 
     with app.app_context():
