--- database/models.py
@@ -1,56 +1,0 @@
-from database.database import db
-from flask_login import UserMixin
-
-
-class User(db.Model, UserMixin):
-    __tablename__ = 'user_'
-
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(250), unique=True, nullable=False)
-    password = db.Column(db.String(250), nullable=False)
-
-
-class Ingredient(db.Model):
-    __tablename__ = 'ingredient'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(100), nullable=False)
-
-
-class IngredientInformation(db.Model):
-    __tablename__ = 'ingredient_information'
-
-    id = db.Column(db.Integer, primary_key=True)
-    ingredient_id = db.Column(db.Integer, db.ForeignKey('ingredient.id'), nullable=False)
-    calories = db.Column(db.Float, nullable=False)
-    proteins = db.Column(db.Float, nullable=False)
-    fats = db.Column(db.Float, nullable=False)
-    carbohydrates = db.Column(db.Float, nullable=False)
-
-
-class Recipe(db.Model):
-    __tablename__ = 'recipe'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(100), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('user_.id'), nullable=False)
-    user = db.relationship('User', backref='recipes')
-
-
-class RecipeIngredient(db.Model):
-    __tablename__ = 'recipe_ingredient'
-
-    id = db.Column(db.Integer, primary_key=True)
-    recipe_id = db.Column(db.Integer, db.ForeignKey('recipe.id'), nullable=False)
-    ingredient_id = db.Column(db.Integer, db.ForeignKey('ingredient.id'), nullable=False)
-    quantity = db.Column(db.Float, nullable=False)
-    unit = db.Column(db.String(20), nullable=False)
-
-
-class RecipeInstruction(db.Model):
-    __tablename__ = 'recipe_instruction'
-
-    id = db.Column(db.Integer, primary_key=True)
-    recipe_id = db.Column(db.Integer, db.ForeignKey('recipe.id'), nullable=False)
-    step_number = db.Column(db.Integer, nullable=False)
-    instruction_text = db.Column(db.Text, nullable=False)
