+++ models/recipe.py
@@ -1,0 +1,29 @@
+from models.database import db
+
+
+class Recipe(db.Model):
+    __tablename__ = 'recipe'
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(100), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('user_.id'), nullable=False)
+    user = db.relationship('User', backref='recipes')
+
+
+class RecipeIngredient(db.Model):
+    __tablename__ = 'recipe_ingredient'
+
+    id = db.Column(db.Integer, primary_key=True)
+    recipe_id = db.Column(db.Integer, db.ForeignKey('recipe.id'), nullable=False)
+    ingredient_id = db.Column(db.Integer, db.ForeignKey('ingredient.id'), nullable=False)
+    quantity = db.Column(db.Float, nullable=False)
+    unit = db.Column(db.String(20), nullable=False)
+
+
+class RecipeInstruction(db.Model):
+    __tablename__ = 'recipe_instruction'
+
+    id = db.Column(db.Integer, primary_key=True)
+    recipe_id = db.Column(db.Integer, db.ForeignKey('recipe.id'), nullable=False)
+    step_number = db.Column(db.Integer, nullable=False)
+    instruction_text = db.Column(db.Text, nullable=False)
