+++ routes/routes.py
@@ -1,0 +1,10 @@
+
+from flask import Blueprint
+from controllers.user_controller import register, login, logout, home
+
+routes = Blueprint('routes', __name__)
+
+routes.route('/register', methods=["GET", "POST"])(register)
+routes.route("/login", methods=["GET", "POST"])(login)
+routes.route("/logout")(logout)
+routes.route("/")(home)
